refactor: extract leaf components from feedback detail route
feedback/$slug/index.tsx was 2364 lines. Extract the self-contained,
props-only leaf components (no shared parent state) into colocated files:

- -types.ts                              shared row/data types
- -components/github-connection-dialog.tsx  GitHub connect dialog + badges
- -components/feedback-event-item.tsx        timeline event row + helpers

index.tsx: 2364 -> 1723 lines. The core stateful FeedbackDetailRoute stays.
routeTree.gen.ts unchanged (-components / -types are excluded from routing).

Verified: tsc clean, 119/119 tests pass, prettier clean.
