+++ src/routes/@{$org}/$project/feedback/$slug/-types.ts
@@ -1,0 +1,56 @@
+export type ProfileSummary = {
+  id?: string
+  imageUrl?: string | null
+  name?: string | null
+  username: string
+}
+
+export type FeedbackCommentData = {
+  author: ProfileSummary | null
+  content: string
+  createdAt: number | string | Date
+  emoteCounts?: Record<string, { authorProfileIds: string[]; count: number }>
+  id: string
+  initial?: boolean
+  isTeamMember?: boolean
+  updatedTime?: number | string | Date | null
+}
+
+export type FeedbackEventData = {
+  actor?: ProfileSummary | null
+  createdAt: number | string | Date
+  eventType:
+    | "answer_marked"
+    | "answer_unmarked"
+    | "assigned"
+    | "board_changed"
+    | "status_changed"
+    | "title_changed"
+    | "unassigned"
+    | string
+  id: string
+  metadata?: {
+    newValue?: string | null
+    oldValue?: string | null
+  } | null
+  targetProfile?: ProfileSummary | null
+}
+
+export type GitHubConnectionData = {
+  githubNumber: number
+  id: string
+  kind: "issue"
+  state: string
+  title: string
+  url: string
+}
+
+export type GitHubTargetData = {
+  databaseId?: number
+  kind: "issue"
+  nodeId: string
+  number: number
+  state: string
+  title: string
+  url: string
+}
