+++ src/routes/@{$org}/$project/feedback/$slug/index.tsx
@@ -1,32 +1,24 @@
 import {
-  ArrowRightLeft,
   Bell,
   Calendar as CalendarIcon,
   Check,
-  CheckCircle2,
   ChevronDown,
   ChevronRight,
-  CircleX,
   ExternalLink,
-  FolderInput,
   GitBranch,
   Info,
   Link as LinkIcon,
   MessageSquare,
   MoreHorizontal,
   Plus,
-  Search,
   Tag,
   Trash2,
-  UserMinus,
-  UserPlus,
   Users,
 } from "lucide-react"
 import {
   type CSSProperties,
   type FormEvent,
   type KeyboardEvent,
-  memo,
   useEffect,
   useMemo,
   useRef,
@@ -95,17 +87,12 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip"
-import { EditIcon, GithubIcon, StatusIcon } from "@/icons"
+import { EditIcon, StatusIcon } from "@/icons"
 import { useCRPC } from "@/lib/convex/crpc"
 import { crpcServer } from "@/lib/convex/crpc-server"
 import { useSidebarState } from "@/lib/hooks/use-sidebar-state"
 import { cn } from "@/lib/utils"
-import {
-  formatFullDate,
-  formatRelativeDay,
-  formatTimestamp,
-  toTimestamp,
-} from "@/lib/utils/format-timestamp"
+import { formatTimestamp, toTimestamp } from "@/lib/utils/format-timestamp"
 import { FORM_LIMITS } from "@/lib/validation"
 
 import { UpvoteButton } from "../-components/upvote-button"
@@ -116,6 +103,18 @@ import {
   type ThreadComment,
 } from "../../-components/comment-thread"
 import { projectTitle, titleFromSlug, titleMeta } from "@/lib/seo"
+import type {
+  FeedbackCommentData,
+  FeedbackEventData,
+  GitHubConnectionData,
+  ProfileSummary,
+} from "./-types"
+import { FeedbackEventItem } from "./-components/feedback-event-item"
+import {
+  GitHubConnectionDialog,
+  GithubConnectionIcon,
+  GithubIssueStateBadge,
+} from "./-components/github-connection-dialog"
 
 const SIDEBAR_STORAGE_KEY = "feedback-detail-sidebar-state"
 
@@ -202,63 +201,6 @@ function parseMonthTarget(target: string) {
     month: parsed ? pad2(parsed.month) : pad2(now.getMonth() + 1),
     year: parsed ? String(parsed.year) : String(now.getFullYear()),
   }
-}
-
-type ProfileSummary = {
-  id?: string
-  imageUrl?: string | null
-  name?: string | null
-  username: string
-}
-
-type FeedbackCommentData = {
-  author: ProfileSummary | null
-  content: string
-  createdAt: number | string | Date
-  emoteCounts?: Record<string, { authorProfileIds: string[]; count: number }>
-  id: string
-  initial?: boolean
-  isTeamMember?: boolean
-  updatedTime?: number | string | Date | null
-}
-
-type FeedbackEventData = {
-  actor?: ProfileSummary | null
-  createdAt: number | string | Date
-  eventType:
-    | "answer_marked"
-    | "answer_unmarked"
-    | "assigned"
-    | "board_changed"
-    | "status_changed"
-    | "title_changed"
-    | "unassigned"
-    | string
-  id: string
-  metadata?: {
-    newValue?: string | null
-    oldValue?: string | null
-  } | null
-  targetProfile?: ProfileSummary | null
-}
-
-type GitHubConnectionData = {
-  githubNumber: number
-  id: string
-  kind: "issue"
-  state: string
-  title: string
-  url: string
-}
-
-type GitHubTargetData = {
-  databaseId?: number
-  kind: "issue"
-  nodeId: string
-  number: number
-  state: string
-  title: string
-  url: string
 }
 
 export const Route = createFileRoute("/@{$org}/$project/feedback/$slug/")({
@@ -1777,588 +1719,5 @@ function InlineFeedbackTitleEditor({
         </>
       ) : null}
     </div>
-  )
-}
-
-function GithubConnectionIcon() {
-  return <GithubIcon className="size-3.5 shrink-0 text-muted-foreground" />
-}
-
-function GithubIssueStateBadge({ state }: { state: string }) {
-  const normalizedState = state.trim().toLowerCase()
-  const isOpen = normalizedState === "open"
-
-  return (
-    <span
-      className={cn(
-        "shrink-0 rounded-full px-1.5 py-0.5 text-[10px] leading-none font-medium whitespace-nowrap capitalize",
-        isOpen
-          ? "bg-green-500/10 text-green-600 dark:text-green-400"
-          : "bg-muted text-muted-foreground"
-      )}
-    >
-      {normalizedState || "unknown"}
-    </span>
-  )
-}
-
-function GitHubConnectionDialog({
-  feedbackId,
-  feedbackTitle,
-  onOpenChange,
-  open,
-  orgSlug,
-  projectSlug,
-}: {
-  feedbackId: string
-  feedbackTitle: string
-  onOpenChange: (open: boolean) => void
-  open: boolean
-  orgSlug: string
-  projectSlug: string
-}) {
-  const crpc = useCRPC()
-  const [mode, setMode] = useState<"existing" | "create">("existing")
-  const [query, setQuery] = useState("")
-  const [selectedTarget, setSelectedTarget] = useState<GitHubTargetData | null>(
-    null
-  )
-  const [title, setTitle] = useState("")
-  const [body, setBody] = useState("")
-  const [localError, setLocalError] = useState("")
-
-  const searchMutation = useMutation(
-    crpc.feedbackGithub.searchTargets.mutationOptions()
-  )
-  const availabilityQuery = useQuery(
-    crpc.feedbackGithub.getAvailability.queryOptions(
-      { feedbackId },
-      { enabled: open, skipUnauth: true }
-    )
-  )
-  const connectExistingMutation = useMutation(
-    // The connections list is a live subscription, so it updates on its own
-    // once the connection is written server-side — just close the dialog.
-    crpc.feedbackGithub.connectExisting.mutationOptions({
-      onSuccess: () => {
-        onOpenChange(false)
-      },
-    })
-  )
-  const createMutation = useMutation(
-    crpc.feedbackGithub.createAndConnect.mutationOptions({
-      onSuccess: () => {
-        onOpenChange(false)
-      },
-    })
-  )
-
-  useEffect(() => {
-    if (!open) {
-      setMode("existing")
-      setQuery("")
-      setSelectedTarget(null)
-      setTitle("")
-      setBody("")
-      setLocalError("")
-      searchMutation.reset()
-      connectExistingMutation.reset()
-      createMutation.reset()
-    }
-  }, [open])
-
-  useEffect(() => {
-    if (!open || mode !== "existing") return
-    if (availabilityQuery.data && !availabilityQuery.data.issuesEnabled) return
-
-    const timeout = window.setTimeout(() => {
-      searchMutation.mutate({
-        feedbackId,
-        kind: "issue",
-        query: query.slice(0, FORM_LIMITS.feedbackSearch),
-      })
-      setSelectedTarget(null)
-    }, 300)
-    return () => window.clearTimeout(timeout)
-  }, [availabilityQuery.data?.issuesEnabled, feedbackId, mode, open, query])
-
-  const searchResults = (searchMutation.data ?? []) as GitHubTargetData[]
-  const availability = availabilityQuery.data
-  const sourceDisabled = !!availability && !availability.issuesEnabled
-  const writeDisabled =
-    !!availability && availability.connected && !availability.writable
-  const repoMissing = !!availability && !availability.connected
-  const error =
-    localError ||
-    (availabilityQuery.error?.message ??
-      connectExistingMutation.error?.message ??
-      createMutation.error?.message ??
-      searchMutation.error?.message)
-  const feedbackUrl =
-    typeof window === "undefined" ? "" : window.location.href.split("#")[0]
-  const canCreate =
-    title.trim().length > 0 &&
-    title.trim().length <= FORM_LIMITS.githubTitle &&
-    body.trim().length <= FORM_LIMITS.githubBody &&
-    !createMutation.isPending &&
-    !sourceDisabled &&
-    !writeDisabled &&
-    !repoMissing
-  const canConnect =
-    !!selectedTarget &&
-    !sourceDisabled &&
-    !writeDisabled &&
-    !repoMissing &&
-    !connectExistingMutation.isPending &&
-    !searchMutation.isPending
-
-  function handleConnectExisting() {
-    setLocalError("")
-    if (!selectedTarget || !feedbackUrl) return
-
-    connectExistingMutation.mutate({
-      feedbackId,
-      feedbackUrl,
-      githubNumber: selectedTarget.number,
-      kind: "issue",
-    })
-  }
-
-  function handleCreate() {
-    setLocalError("")
-    if (!feedbackUrl) return
-    if (title.trim().length > FORM_LIMITS.githubTitle) {
-      setLocalError(
-        `GitHub issue titles must be ${FORM_LIMITS.githubTitle} characters or fewer.`
-      )
-      return
-    }
-    if (body.trim().length > FORM_LIMITS.githubBody) {
-      setLocalError(
-        `GitHub issue bodies must be ${FORM_LIMITS.githubBody} characters or fewer.`
-      )
-      return
-    }
-
-    createMutation.mutate({
-      body,
-      feedbackId,
-      feedbackUrl,
-      kind: "issue",
-      title: title.trim(),
-    })
-  }
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="flex max-h-[92vh] flex-col gap-0 overflow-hidden p-0 sm:max-w-xl">
-        {/* Header */}
-        <div className="flex items-start justify-between px-5 pt-5 pb-4">
-          <div className="flex items-start gap-3">
-            <div className="mt-0.5 flex size-8 shrink-0 items-center justify-center rounded-lg border bg-muted/50">
-              <GithubIcon className="size-4 text-muted-foreground" />
-            </div>
-            <div>
-              <h2 className="text-sm leading-tight font-semibold">
-                Connect GitHub
-              </h2>
-              <p className="mt-0.5 line-clamp-1 max-w-xs text-xs text-muted-foreground">
-                {feedbackTitle}
-              </p>
-            </div>
-          </div>
-        </div>
-
-        {/* Controls */}
-        <div className="flex items-center gap-2 border-y bg-muted/30 px-5 py-2">
-          {/* Mode segmented control */}
-          <div className="flex items-center rounded-md bg-background p-0.5 shadow-sm ring-1 ring-border/60">
-            <button
-              className={cn(
-                "rounded px-3 py-1 text-xs font-medium transition-all",
-                mode === "existing"
-                  ? "bg-primary text-primary-foreground shadow-sm"
-                  : "text-muted-foreground hover:text-foreground"
-              )}
-              onClick={() => setMode("existing")}
-              type="button"
-            >
-              Link existing
-            </button>
-            <button
-              className={cn(
-                "rounded px-3 py-1 text-xs font-medium transition-all",
-                mode === "create"
-                  ? "bg-primary text-primary-foreground shadow-sm"
-                  : "text-muted-foreground hover:text-foreground"
-              )}
-              onClick={() => setMode("create")}
-              type="button"
-            >
-              Create new
-            </button>
-          </div>
-        </div>
-
-        {/* Body */}
-        <div className="min-h-0 flex-1 overflow-y-auto px-5 py-4">
-          {availabilityQuery.isLoading ? (
-            <div className="space-y-2">
-              <div className="h-9 animate-pulse rounded-lg bg-muted/50" />
-              <div className="h-14 animate-pulse rounded-lg bg-muted/50" />
-              <div className="h-14 animate-pulse rounded-lg bg-muted/50" />
-            </div>
-          ) : sourceDisabled ? (
-            <GitHubConnectionNotice
-              description="Issues are not enabled for this project's connected GitHub repository."
-              orgSlug={orgSlug}
-              projectSlug={projectSlug}
-              title="Enable GitHub Issues"
-            />
-          ) : repoMissing ? (
-            <GitHubConnectionNotice
-              description="Connect a GitHub repository before linking feedback to issues."
-              orgSlug={orgSlug}
-              projectSlug={projectSlug}
-              title="Connect a GitHub repository"
-            />
-          ) : writeDisabled ? (
-            <GitHubConnectionNotice
-              description="This GitHub repository is connected read-only. Reconnect it with read/write access before linking feedback."
-              orgSlug={orgSlug}
-              projectSlug={projectSlug}
-              title="Reconnect with write access"
-            />
-          ) : mode === "existing" ? (
-            <div className="space-y-3">
-              <div className="relative">
-                <Search className="absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-muted-foreground" />
-                <Input
-                  className="h-9 pl-9 text-sm"
-                  maxLength={FORM_LIMITS.feedbackSearch}
-                  onChange={(event) => setQuery(event.target.value)}
-                  placeholder="Search issues..."
-                  value={query}
-                />
-              </div>
-              <div className="space-y-1">
-                {searchMutation.isPending ? (
-                  <div className="space-y-1.5">
-                    <div className="h-14 animate-pulse rounded-lg bg-muted/50" />
-                    <div className="h-14 animate-pulse rounded-lg bg-muted/50" />
-                  </div>
-                ) : searchResults.length > 0 ? (
-                  searchResults.map((target) => (
-                    <button
-                      className={cn(
-                        "flex w-full min-w-0 items-center gap-3 rounded-lg border border-transparent px-3 py-2.5 text-left transition-all hover:bg-muted/50",
-                        selectedTarget?.nodeId === target.nodeId
-                          ? "border-primary/25 bg-primary/5 ring-1 ring-primary/20"
-                          : "hover:border-border/60"
-                      )}
-                      key={target.nodeId}
-                      onClick={() => setSelectedTarget(target)}
-                      type="button"
-                    >
-                      <div className="flex size-7 shrink-0 items-center justify-center rounded-md bg-muted">
-                        <GithubConnectionIcon />
-                      </div>
-                      <span className="min-w-0 flex-1">
-                        <span className="block truncate text-sm leading-tight font-medium">
-                          #{target.number} {target.title}
-                        </span>
-                        <span className="mt-0.5 flex items-center gap-1.5 text-xs text-muted-foreground">
-                          <span
-                            className={cn(
-                              "size-1.5 rounded-full",
-                              target.state === "open"
-                                ? "bg-green-500"
-                                : "bg-muted-foreground/50"
-                            )}
-                          />
-                          {target.state}
-                        </span>
-                      </span>
-                      {selectedTarget?.nodeId === target.nodeId ? (
-                        <Check className="size-3.5 shrink-0 text-primary" />
-                      ) : null}
-                    </button>
-                  ))
-                ) : (
-                  <div className="rounded-lg border border-dashed py-8 text-center text-sm text-muted-foreground">
-                    {searchMutation.data
-                      ? "No issues found."
-                      : "Loading issues..."}
-                  </div>
-                )}
-              </div>
-            </div>
-          ) : (
-            <div className="space-y-3">
-              <div className="space-y-1.5">
-                <label className="text-xs font-medium text-muted-foreground">
-                  Title
-                </label>
-                <Input
-                  className="h-9 text-sm"
-                  maxLength={FORM_LIMITS.githubTitle}
-                  onChange={(event) => setTitle(event.target.value)}
-                  placeholder="Issue title"
-                  value={title}
-                />
-              </div>
-              <div className="space-y-1.5">
-                <label className="text-xs font-medium text-muted-foreground">
-                  Body{" "}
-                  <span className="font-normal text-muted-foreground/60">
-                    (optional)
-                  </span>
-                </label>
-                <Textarea
-                  className="min-h-28 resize-none text-sm"
-                  maxLength={FORM_LIMITS.githubBody}
-                  onChange={(event) => setBody(event.target.value)}
-                  placeholder="Add a description…"
-                  value={body}
-                />
-              </div>
-            </div>
-          )}
-        </div>
-
-        {/* Footer */}
-        <div className="flex items-center justify-between border-t bg-muted/20 px-5 py-3">
-          {error ? (
-            <p className="text-xs text-destructive">{error}</p>
-          ) : (
-            <span />
-          )}
-          <div className="flex items-center gap-2">
-            <Button
-              onClick={() => onOpenChange(false)}
-              size="sm"
-              type="button"
-              variant="outline"
-            >
-              Cancel
-            </Button>
-            {mode === "existing" ? (
-              <Button
-                disabled={!canConnect}
-                onClick={handleConnectExisting}
-                size="sm"
-                type="button"
-              >
-                <LinkIcon className="size-3.5" />
-                Connect
-              </Button>
-            ) : (
-              <Button
-                disabled={!canCreate}
-                onClick={handleCreate}
-                size="sm"
-                type="button"
-              >
-                <Plus className="size-3.5" />
-                Create & connect
-              </Button>
-            )}
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
-  )
-}
-
-function GitHubConnectionNotice({
-  description,
-  orgSlug,
-  projectSlug,
-  title,
-}: {
-  description: string
-  orgSlug: string
-  projectSlug: string
-  title: string
-}) {
-  return (
-    <div className="rounded-lg border border-dashed p-5">
-      <div className="flex items-start gap-3">
-        <div className="flex size-8 shrink-0 items-center justify-center rounded-md bg-muted">
-          <GithubIcon className="size-3.5 text-muted-foreground" />
-        </div>
-        <div className="min-w-0 flex-1 space-y-3">
-          <div>
-            <h3 className="text-sm leading-tight font-medium">{title}</h3>
-            <p className="mt-1 text-xs text-muted-foreground">{description}</p>
-          </div>
-          <Button asChild size="sm" type="button" variant="outline">
-            <Link
-              params={{ org: orgSlug, project: projectSlug }}
-              to="/@{$org}/$project/settings/integrations"
-            >
-              <GithubIcon className="size-3.5" />
-              Open GitHub settings
-            </Link>
-          </Button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-// Memoized: timeline event rows take only the stable `event` prop, so they
-// skip re-rendering when unrelated top-level state (dialogs, sheets) changes.
-const FeedbackEventItem = memo(function FeedbackEventItem({
-  event,
-}: {
-  event: FeedbackEventData
-}) {
-  const Icon = getEventIcon(event.eventType)
-  const createdAt = toTimestamp(event.createdAt)
-
-  return (
-    <li className="relative z-10 flex items-start gap-3 py-2 pl-4">
-      <div className="ml-0.5 flex h-6 w-8 shrink-0 items-center justify-center">
-        <div className="flex h-6 w-6 items-center justify-center rounded-full border bg-muted">
-          <Icon className="h-3 w-3 text-muted-foreground" />
-        </div>
-      </div>
-      <div className="flex min-w-0 flex-1 flex-wrap items-center gap-x-1 text-sm text-muted-foreground">
-        {event.actor ? (
-          <Link
-            className="font-medium text-foreground hocus:underline"
-            params={{ username: event.actor.username }}
-            to="/u/$username"
-          >
-            @{event.actor.username}
-          </Link>
-        ) : (
-          <span className="font-medium text-foreground">Someone</span>
-        )}{" "}
-        {getEventDescription(event)}{" "}
-        <Tooltip>
-          <TooltipTrigger asChild delay={100}>
-            <span
-              className="cursor-pointer border-b border-dotted border-foreground/30 text-foreground/50"
-              suppressHydrationWarning
-            >
-              {formatRelativeDay(createdAt)}
-            </span>
-          </TooltipTrigger>
-          <TooltipContent>
-            <span suppressHydrationWarning>{formatFullDate(createdAt)}</span>
-          </TooltipContent>
-        </Tooltip>
-      </div>
-    </li>
-  )
-})
-
-function getEventIcon(eventType: FeedbackEventData["eventType"]) {
-  switch (eventType) {
-    case "status_changed":
-      return ArrowRightLeft
-    case "board_changed":
-      return FolderInput
-    case "title_changed":
-      return Tag
-    case "assigned":
-      return UserPlus
-    case "unassigned":
-      return UserMinus
-    case "answer_marked":
-      return CheckCircle2
-    case "answer_unmarked":
-      return CircleX
-    default:
-      return Check
-  }
-}
-
-function getEventDescription(event: FeedbackEventData) {
-  const { eventType, metadata, targetProfile } = event
-
-  switch (eventType) {
-    case "status_changed":
-      return (
-        <span className="inline-flex flex-wrap items-center gap-1">
-          changed status from <StatusPill status={metadata?.oldValue} /> to{" "}
-          <StatusPill status={metadata?.newValue} />
-        </span>
-      )
-    case "board_changed":
-      return (
-        <span>
-          moved to board{" "}
-          <span className="rounded bg-muted px-1.5 py-0.5 text-xs font-medium">
-            {metadata?.newValue ?? "Unknown"}
-          </span>
-        </span>
-      )
-    case "title_changed":
-      return (
-        <span>
-          changed title from{" "}
-          <span className="rounded bg-muted px-1.5 py-0.5 text-xs font-medium">
-            {metadata?.oldValue ?? "Untitled"}
-          </span>{" "}
-          to{" "}
-          <span className="rounded bg-muted px-1.5 py-0.5 text-xs font-medium">
-            {metadata?.newValue ?? "Untitled"}
-          </span>
-        </span>
-      )
-    case "assigned":
-      return (
-        <span>
-          assigned{" "}
-          {targetProfile ? (
-            <Link
-              className="font-medium hocus:underline"
-              params={{ username: targetProfile.username }}
-              to="/u/$username"
-            >
-              @{targetProfile.username}
-            </Link>
-          ) : (
-            <span className="text-muted-foreground">unknown user</span>
-          )}
-        </span>
-      )
-    case "unassigned":
-      return (
-        <span>
-          unassigned{" "}
-          {targetProfile ? (
-            <Link
-              className="font-medium hocus:underline"
-              params={{ username: targetProfile.username }}
-              to="/u/$username"
-            >
-              @{targetProfile.username}
-            </Link>
-          ) : (
-            <span className="text-muted-foreground">unknown user</span>
-          )}
-        </span>
-      )
-    case "answer_marked":
-      return <span>marked a comment as the answer</span>
-    case "answer_unmarked":
-      return <span>unmarked the answer</span>
-    default:
-      return <span>made a change</span>
-  }
-}
-
-function StatusPill({ status }: { status?: string | null }) {
-  return (
-    <span className="inline-flex items-center gap-1 rounded bg-muted px-1.5 py-0.5">
-      {status ? (
-        <StatusIcon colored size="12" status={status as never} />
-      ) : null}
-      <span className="text-xs font-medium">{status ?? "none"}</span>
-    </span>
   )
 }
